+++ tests/Transformers/TextNormalizerTest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Rubix\ML\Tests\Transformers;
+
+use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Transformers\Transformer;
+use Rubix\ML\Transformers\TextNormalizer;
+use PHPUnit\Framework\TestCase;
+
+class TextNormalizerTest extends TestCase
+{
+    protected $transformer;
+
+    protected $dataset;
+
+    public function setUp()
+    {
+        $this->dataset = Unlabeled::quick([
+            ['The quick brown fox jumped over the lazy man sitting at a bus'
+                . ' stop drinking a can of Coke'],
+            ['with a Dandy   ubrella'],
+        ]);
+
+        $this->transformer = new TextNormalizer();
+    }
+
+    public function test_build_transformer()
+    {
+        $this->assertInstanceOf(TextNormalizer::class, $this->transformer);
+        $this->assertInstanceOf(Transformer::class, $this->transformer);
+    }
+
+    public function test_transform()
+    {
+        $this->dataset->apply($this->transformer);
+    
+        $outcome = [
+            ['the quick brown fox jumped over the lazy man sitting at a bus'
+                . ' stop drinking a can of coke'],
+            ['with a dandy ubrella'],
+        ];
+    
+        $this->assertEquals($outcome, $this->dataset->samples());
+    }
+}